refactor(core-rs): split run_command process plumbing into actions/shell (T6.3)
run_command.rs crossed the 300-line rule; the shell resolution, capped
stream capture, and spawn move to actions/shell.rs unchanged.
